requirements.txt
flask
strawberry-graphql
sqlalchemy
